# netflix/src/api/swagger/api.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Netflix API
  description: Netflix API
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /v1
tags:
  - name: public
    description: Public operations (authentication is not required)

  - name: user
    description: User operations (user authentication is required)

  - name: admin
    description: Admin operations (admin authentication is required)

paths:
  /user:
    post:
      description: Create a new user
      tags:
        - public
      operationId: createUser
      x-eov-operation-handler: api/controllers/netflix.controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Account has been successfully created

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error

  /user/login:
    post:
      description: User login
      tags:
        - public
      operationId: login
      x-eov-operation-handler: api/controllers/netflix.controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: integer
                    format: uuid
                example:
                  sessionId: 123e4567-e89b-12d3-a456-426614174000

        "400":
          description: Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid username or password, please double-check your login details and try again.

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error
       
  /user/logout:
    post:
      description: User logout
      tags:
        - user
      operationId: logout
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
          - in: header
            name: X-Session-ID
            description: The user must have an authentication ID for account operations.
            required: true
            schema:
              type: string
              format: uuid
      
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Logout successfully

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error    
  
  /queue:
    get:
      description: List movies queue rented by the user
      tags:
        - user
      operationId: queue
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
          - in: header
            name: X-Session-ID
            description: The user must have an authentication ID for account operations.
            required: true
            schema:
              type: string
              format: uuid

          - in: query
            name: order
            description: The user can order the list by ID.
            required: true
            schema:
              type: boolean
              default: false

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                 $ref: "#/components/schemas/MovieArray"

        "400":
          description: Invalid session ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Please login!

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error
      
  /movies:
    get:
      description: Get movie by title or id
      tags:
        - user
      operationId: findMovie
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
        - in: header
          name: X-Session-ID
          description: The user must have an authentication ID for account operations
          required: true
          schema:
            type: string
            format: uuid

        - in: query
          name: title
          description: The title of the movie.
          required: false
          schema:
            type: string

        - in: query
          name: id
          description: The ID of the movie.
          required: false
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieArray"

        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: No movies found!

        "400":
          description: Invalid session ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Please login!

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error

    post:
      description: Create movie
      tags:
        - admin
      operationId: createMovie
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
        - in: header
          name: X-Admin-API-key
          description: The user must have an admin-api-key for admin operations.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"

      responses:
          "201":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Movie"

          "400":
            description: Invalid API Key
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                  example:
                    message: Invalid admin key!

          "500":
            description: Unexpected error
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                  example:
                    message: Unexpected error

    delete:
      description: Delete movie
      tags:
        - admin
      operationId: deleteMovie
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
        - in: header
          name: X-Admin-API-key
          description: The user must have an admin-api-key for admin operations.
          required: true
          schema:
            type: string
            format: uuid
        
        - in: query
          name: id
          description: The ID of the movie you want to delete.
          required: true
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: You deleted the movie successfully.

        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: No movies found with this id.

        "400":
          description: Invalid API Key
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid admin key!

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error

    patch:
      description: Update movie
      tags:
        - admin
      operationId: updateMovie
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
        - in: header
          name: X-Admin-API-key
          description: The user must have an admin-api-key for admin operations.
          required: true
          schema:
            type: string
            format: uuid

        - in: query
          name: id
          description: The ID of the movie.
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"

        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: No movies found with this id.

        "400":
          description: Invalid API Key
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid admin key!

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error

  /movies/{movieId}/rent:
    get:
      description: Rent a movie by ID
      tags:
        - user
      operationId: rentMovieById
      x-eov-operation-handler: api/controllers/netflix.controller
      parameters:
        - in: header
          name: X-Session-ID
          description: The user must have an authentication ID for account operations.
          required: true
          schema:
            type: string
            format: uuid

        - in: path
          name: movieId
          description: The ID of the movie.
          required: true
          schema:
            type: string

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: You rented the movie successfully.

        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: No movies found with this id.

        "400":
          description: Invalid session ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Please login!

        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Unexpected error

components:
  schemas:
    User:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: The username
        password:
          type: string
          description: The password
      example:
        username: roland
        password: password123

    Movie:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        directed-by:
          type: string
        actors:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
        type:
          type: string
          enum:
            - TV Show
            - Movie
        id:
          type: integer
          readOnly: true

      required:
            - id
            - title
            - description
            - type

      example:
        title: "Transformers"
        description: "Transformers (2007) is a high-octane science fiction action film directed by Michael Bay, focusing on a war between two factions of alien robots—the heroic Autobots and evil Decepticons—who arrive on Earth searching for the AllSpark, a powerful artifact."
        directed-by: Michael Bay
        actors: [Shia LaBeouf, Megan Fox]
        categories: [sci-fi, action]
        type: Movie
        id: 3

    MovieArray:
      type: array
      items:
        $ref: "#/components/schemas/Movie"